+++ order/src/main/java/com/toanyone/order/infrastructure/converter/OrderMessageConverter.java
@@ -1,4 +1,4 @@
-package com.toanyone.order.application.mapper;
+package com.toanyone.order.infrastructure.converter;
 
 import com.toanyone.order.application.dto.request.OrderCreateServiceDto;
 import com.toanyone.order.message.DeliveryRequestMessage;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 import java.util.stream.Collectors;
 
 @Component
-public class MessageConverter {
+public class OrderMessageConverter {
 
     public DeliveryRequestMessage toOrderDeliveryMessage(OrderCreateServiceDto orderCreateRequest, Long orderId, Long arrivalHubId, Long departureHubId) {
