+++ order/src/main/java/com/toanyone/order/presentation/mapper/OrderApplicationToServiceDtoMapper.java
@@ -4,7 +4,6 @@ import com.toanyone.order.application.dto.request.OrderCancelServiceDto;
 import com.toanyone.order.application.dto.request.OrderCreateServiceDto;
 import com.toanyone.order.application.dto.request.OrderFindAllCondition;
 import com.toanyone.order.application.dto.request.OrderSearchCondition;
-import com.toanyone.order.presentation.dto.request.OrderCancelRequestDto;
 import com.toanyone.order.presentation.dto.request.OrderCreateRequestDto;
 import com.toanyone.order.presentation.dto.request.OrderFindAllRequestDto;
 import com.toanyone.order.presentation.dto.request.OrderSearchRequestDto;
@@ -14,7 +13,7 @@ import java.util.stream.Collectors;
 
 
 @Component
-public class OrderMapper {
+public class OrderApplicationToServiceDtoMapper {
 
     public OrderCreateServiceDto toOrderCreateServiceDto(OrderCreateRequestDto request) {
         return OrderCreateServiceDto.builder()
